round 9: full audit remediation + legacy removal — 25 files fixed across 6 domains
CRITICAL (6): infrastructure port alignment to docs (13830/13810/13820),
Redis healthcheck with password auth, REDIS_URL with credentials,
Caddyfile rewritten for v3, Caddy reverse proxy added to compose,
Seed script seed_v3.py created (8-step baseline data)

HIGH (11): isLoggedIn SSR guard, useEffect deps/cleanup in admin pages,
staff listing server-side filtering, GalleryUpload stable keys,
E2E weak assertions tightened (200 vs 400/422), cancel rejection == 409,
brute-force endpoint guessing removed, NEXT_PUBLIC_API_URL relative paths,
CORS_ORIGINS env-driven, staff requestRaw abort check, staff error/loading.tsx

MEDIUM (12): 13 empty model dirs removed, admin/endpoints __init__.py,
services re-exports, PwaTranslationsTab /admin/ prefix, useAudienceSegments
cleanup, CSS drawer conflict removed, duplicate CSS selectors cleaned,
config.py cors_origins aligned, .env.example fixed, E2E BASE_URL env-var

LEGACY: ~1.1GB removed — old backend/, frontend/, customer-app/,
docker-compose.yml, .env/.env.local, scripts/seed/, scripts/fnb-manage.sh,
third-party mock scripts, __pycache__ directories

+++ v3/backend/app/models/cart.py
@@ -1,6 +1,7 @@
 """Cart & Checkout models."""
 
 from datetime import datetime, timezone
+from decimal import Decimal
 from typing import List
 
 from sqlalchemy import (
@@ -31,7 +32,7 @@ class CustomerCart(Base, TimestampMixin):
         ForeignKey("stores.id", ondelete="CASCADE"), nullable=False, index=True
     )
     item_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    subtotal: Mapped[float] = mapped_column(Numeric(10, 4), nullable=False, default=0)
+    subtotal: Mapped[Decimal] = mapped_column(Numeric(10, 4), nullable=False, default=0)
     last_activity_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True),
         nullable=False,
@@ -65,10 +66,10 @@ class CartLineItem(Base):
         ForeignKey("menu_variants.id", ondelete="SET NULL"), nullable=True, index=True
     )
     quantity: Mapped[int] = mapped_column(Integer, nullable=False)
-    unit_price: Mapped[float] = mapped_column(Numeric(10, 4), nullable=False)
-    line_total: Mapped[float] = mapped_column(Numeric(10, 4), nullable=False)
+    unit_price: Mapped[Decimal] = mapped_column(Numeric(10, 4), nullable=False)
+    line_total: Mapped[Decimal] = mapped_column(Numeric(10, 4), nullable=False)
     selected_modifiers: Mapped[dict] = mapped_column(JSONB, nullable=False, default=dict)
-    modifier_total: Mapped[float] = mapped_column(Numeric(10, 4), nullable=False, default=0)
+    modifier_total: Mapped[Decimal] = mapped_column(Numeric(10, 4), nullable=False, default=0)
     special_instructions: Mapped[str | None] = mapped_column(String(255), nullable=True)
     added_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True),
